backend/requirements.txt
Flask
Flask-SQLAlchemy==2.5.0
flask-praetorian
python-dotenv
flask-restx
flask-cors
flask-marshmallow
marshmallow
marshmallow-sqlalchemy
passlib
Flask-Limiter
pycryptodome
sqlalchemy>=1.4.0